Break internalFunctions into separate files
