+++ src/internal/_clamp.ts
@@ -1,0 +1,3 @@
+export function _clamp(value: number, min: number, max: number) {
+  return value < min ? min : value > max ? max : value;
+}
