+++ src/internal/const.ts
